Move existing files to Desktop.Analyzers
CA2229, CA2235, CA2237
