+++ PortFxCop/src/Desktop.Analyzers/CSharp/CSharpMarkAllNonSerializableFields.Fixer.cs
@@ -1,20 +1,24 @@
-// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
-                                 
+﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
+
 using System.Composition;
-using System.Diagnostics;  
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CodeFixes;
 using Microsoft.CodeAnalysis.CSharp;
-using Microsoft.CodeAnalysis.CSharp.Syntax;
 
 namespace Desktop.Analyzers
-{                                 
-    /// <summary>
-    /// CA2235: Mark all non-serializable fields
-    /// </summary>
-    [ExportCodeFixProvider(LanguageNames.CSharp), Shared]
+{
+    [ExportCodeFixProvider(LanguageNames.CSharp, Name = "CA2237 CodeFix provider"), Shared]
     public class CSharpMarkAllNonSerializableFieldsFixer : MarkAllNonSerializableFieldsFixer
-    { 
+    {
+        protected override SyntaxNode GetFieldDeclarationNode(SyntaxNode node)
+        {
+            var fieldNode = node;
+            while (fieldNode != null && fieldNode.Kind() != SyntaxKind.FieldDeclaration)
+            {
+                fieldNode = fieldNode.Parent;
+            }
 
+            return fieldNode?.Kind() == SyntaxKind.FieldDeclaration ? fieldNode : null;
+        }
     }
 }
