--- PortFxCop/src/Desktop.Analyzers/Core/MarkAllNonSerializableFields.cs
@@ -1,38 +1,0 @@
-// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
-
-using System.Collections.Immutable;
-using Microsoft.CodeAnalysis.Diagnostics;
-using Microsoft.CodeAnalysis;
-
-namespace Desktop.Analyzers
-{                   
-    /// <summary>
-    /// CA2235: Mark all non-serializable fields
-    /// </summary>
-    public abstract class MarkAllNonSerializableFieldsAnalyzer : DiagnosticAnalyzer
-    {
-        internal const string RuleId = "CA2235";
-
-        private static readonly LocalizableString s_localizableTitle = new LocalizableResourceString(nameof(DesktopAnalyzersResources.MarkAllNonSerializableFieldsTitle), DesktopAnalyzersResources.ResourceManager, typeof(DesktopAnalyzersResources));
-        
-        private static readonly LocalizableString s_localizableMessage = new LocalizableResourceString(nameof(DesktopAnalyzersResources.MarkAllNonSerializableFieldsMessage), DesktopAnalyzersResources.ResourceManager, typeof(DesktopAnalyzersResources));
-        private static readonly LocalizableString s_localizableDescription = new LocalizableResourceString(nameof(DesktopAnalyzersResources.MarkAllNonSerializableFieldsDescription), DesktopAnalyzersResources.ResourceManager, typeof(DesktopAnalyzersResources));
-        
-        internal static DiagnosticDescriptor Rule = new DiagnosticDescriptor(RuleId,
-                                                                             s_localizableTitle,
-                                                                             s_localizableMessage,
-                                                                             DiagnosticCategory.Usage,
-                                                                             DiagnosticSeverity.Warning,
-                                                                             isEnabledByDefault: false,
-                                                                             description: s_localizableDescription,
-                                                                             helpLinkUri: null,     // TODO: add MSDN url
-                                                                             customTags: WellKnownDiagnosticTags.Telemetry);
-
-        public override ImmutableArray<DiagnosticDescriptor> SupportedDiagnostics => ImmutableArray.Create(Rule);
-
-        public override void Initialize(AnalysisContext analysisContext)
-        { 
-            
-        }
-    }
-}
