--- PortFxCop/src/Desktop.Analyzers/VisualBasic/BasicImplementSerializationConstructors.Fixer.vb
@@ -1,19 +1,0 @@
-' Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
-
-Imports System.Composition
-Imports System.Linq
-Imports Microsoft.CodeAnalysis
-Imports Microsoft.CodeAnalysis.CodeFixes
-Imports Microsoft.CodeAnalysis.VisualBasic
-Imports Microsoft.CodeAnalysis.VisualBasic.Syntax
-
-Namespace Desktop.Analyzers     
-    ''' <summary>
-    ''' CA2229: Implement serialization constructors
-    ''' </summary>
-    <ExportCodeFixProvider(LanguageNames.VisualBasic), [Shared]>
-    Public NotInheritable Class BasicImplementSerializationConstructorsFixer
-        Inherits ImplementSerializationConstructorsFixer 
-
-    End Class
-End Namespace
