+++ PortFxCop/src/Desktop.Analyzers/VisualBasic/BasicMarkAllNonSerializableFields.Fixer.vb
@@ -1,19 +1,22 @@
-' Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
+﻿' Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
 Imports System.Composition
-Imports System.Linq
 Imports Microsoft.CodeAnalysis
 Imports Microsoft.CodeAnalysis.CodeFixes
 Imports Microsoft.CodeAnalysis.VisualBasic
-Imports Microsoft.CodeAnalysis.VisualBasic.Syntax
 
-Namespace Desktop.Analyzers     
-    ''' <summary>
-    ''' CA2235: Mark all non-serializable fields
-    ''' </summary>
-    <ExportCodeFixProvider(LanguageNames.VisualBasic), [Shared]>
-    Public NotInheritable Class BasicMarkAllNonSerializableFieldsFixer
-        Inherits MarkAllNonSerializableFieldsFixer 
+Namespace Desktop.Analyzers
+    <ExportCodeFixProvider(LanguageNames.VisualBasic, Name:="CA2229 CodeFix provider"), [Shared]>
+    Public Class BasicMarkAllNonSerializableFieldsFixer
+        Inherits MarkAllNonSerializableFieldsFixer
 
+        Protected Overrides Function GetFieldDeclarationNode(node As SyntaxNode) As SyntaxNode
+            Dim fieldNode = node
+            While fieldNode IsNot Nothing AndAlso fieldNode.Kind() <> SyntaxKind.FieldDeclaration
+                fieldNode = fieldNode.Parent
+            End While
+
+            Return If(fieldNode?.Kind() = SyntaxKind.FieldDeclaration, fieldNode, Nothing)
+        End Function
     End Class
 End Namespace
