--- PortFxCop/src/Desktop.Analyzers/VisualBasic/BasicMarkAllNonSerializableFields.vb
@@ -1,17 +1,0 @@
-' Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
-
-Imports Microsoft.CodeAnalysis
-Imports Microsoft.CodeAnalysis.Diagnostics
-Imports Microsoft.CodeAnalysis.VisualBasic
-Imports Microsoft.CodeAnalysis.VisualBasic.Syntax
-
-Namespace Desktop.Analyzers   
-    ''' <summary>
-    ''' CA2235: Mark all non-serializable fields
-    ''' </summary>
-    <DiagnosticAnalyzer(LanguageNames.VisualBasic)>
-    Public NotInheritable Class BasicMarkAllNonSerializableFieldsAnalyzer
-        Inherits MarkAllNonSerializableFieldsAnalyzer
-
-    End Class
-End Namespace
